refactor batch data cache
